various/python-lambda-http/python-http/requirements.txt
aws-cdk-lib==2.43.1
constructs>=10.0.0,<11.0.0
aws-cdk.aws-apigatewayv2-alpha
aws-cdk.aws-apigatewayv2-integrations-alpha
